Cmd rmm implemented (#1)
* Refactor sendMaxMin function to remove unused parameter and improve output formatting

* Enhance sendMaxMin function to format time output and add date_time utility for time conversion

* Add convertTimeToString utility to date_time module for improved time formatting

* Enhance sendMaxMin function to include minimum temperature output with formatted time

* Remove debug print statements from cmd_rmm and update sendMaxMin function to prepare for console output

* Update date format in convertTimeToString function from d-m-Y to d/m/Y

* Change temperature parameter type to float and format output to one decimal place in LCD functions

* Change temperature parameter type to float in LCD task and format output to one decimal place in max/min task

+++ controller/Controllers/LCD/LCD.cpp
@@ -36,7 +36,7 @@ void write_temp_alarm(bool temp_alarm_enabled) {
   lcd.printf(temp_alarm_enabled ? "T" : " ");
 }
 
-void write_temperature(int temperature) {
+void write_temperature(float temperature) {
   lcd.locate(0, 22);
-  lcd.printf("Temp = %d", temperature);
+  lcd.printf("Temp = %.1f", temperature);
 }
